+++ src/array_of.ts
@@ -1,3 +1,0 @@
-export function arrayOf<V>(...args: ReadonlyArray<V>): ReadonlyArray<V> {
-  return Object.freeze(args);
-}