+++ src/is_array.ts
@@ -1,3 +1,0 @@
-export function isArray<T>(value: any): value is Array<T> {
-  return value instanceof Array;
-}