+++ src/keys.ts
@@ -1,3 +1,0 @@
-export function keys(obj: object): ReadonlyArray<string> {
-  return Object.freeze(Object.keys(obj));
-}