+++ src/path.ts
@@ -1,5 +1,3 @@
-import { reduce } from './reduce';
-import { isArray } from './is_array';
 import { isNothing } from './is_nothing';
 import { Maybe, maybeOf } from './maybe';
 
@@ -7,12 +5,11 @@ export function path<R>(
   keysList: string | Array<string>,
   obj: { [index: string]: any }
 ): Maybe<R> {
-  const partsOfPath = isArray(keysList) ? keysList : keysList.split('.');
+  const partsOfPath = Array.isArray(keysList) ? keysList : keysList.split('.');
 
   // @ts-ignore
   return maybeOf(
-    reduce(
-      partsOfPath,
+    partsOfPath.reduce(
       (value, part) =>
         isNothing(value)
           ? null