+++ src/tap.ts
@@ -1,4 +1,0 @@
-export function tap<T>(value: T, fn: (value: T) => any): T {
-  fn(value);
-  return value;
-}