+++ src/wrapper.ts
@@ -2,14 +2,14 @@ import { Monad } from './interfaces/monad';
 import { Comonad } from './interfaces/comonad';
 
 class WrapperConstructor<T> implements Comonad, Monad<T> {
-  private constructor(private readonly value: T) {}
+  private constructor(private readonly _value: T) {}
 
   static wrap<U>(value: U): Wrapper<U> {
     return new WrapperConstructor(value);
   }
 
   map<R>(fn: (value: T) => R): Wrapper<R> {
-    return WrapperConstructor.wrap(fn(this.value));
+    return WrapperConstructor.wrap(fn(this._value));
   }
 
   apply<R>(other: Wrapper<(value: T) => R>): Wrapper<R> {
@@ -17,11 +17,11 @@ class WrapperConstructor<T> implements Comonad, Monad<T> {
   }
 
   chain<R>(fn: (value: T) => Wrapper<R>): Wrapper<R> {
-    return fn(this.value);
+    return fn(this._value);
   }
 
   extract(): T {
-    return this.value;
+    return this._value;
   }
 }