+++ test/reject.test.ts
@@ -1,8 +1,0 @@
-import { reject } from '../src';
-
-describe('reject', () => {
-  test('reject creates rejected promise', () => {
-    expect(reject()).rejects.toBeFalsy();
-    expect(reject(new Error())).rejects.toEqual(new Error());
-  });
-});