+++ test/resolve.test.ts
@@ -1,8 +1,0 @@
-import { resolve } from '../src';
-
-describe('resolve', () => {
-  test('resolve creates resolved promise', () => {
-    expect(resolve()).resolves.toBeFalsy();
-    expect(resolve(5)).resolves.toBe(5);
-  });
-});